requirements.txt
streamlit>=1.36
openai>=1.40.0
pydantic>=2.6
python-dotenv>=1.0
Pillow>=10.3
fpdf2>=2.8
tenacity>=8.2
